[Refactor] #3729 conv_dungeon_feat() をdungeon_type のオブジェクトメソッドへ繰り込んだ
