+++ src/grid/feature.h
@@ -70,4 +70,3 @@ bool is_ascii_graphics(char x);
 FEAT_IDX feat_locked_door_random(int door_type);
 FEAT_IDX feat_jammed_door_random(int door_type);
 void cave_set_feat(PlayerType *player_ptr, POSITION y, POSITION x, FEAT_IDX feat);
-FEAT_IDX conv_dungeon_feat(const FloorType *floor_ptr, FEAT_IDX newfeat);
